+++ test/e2e/_incluster-test-code/memcached_test.go
@@ -65,7 +65,7 @@ func TestMemcached(t *testing.T) {
 	})
 }
 
-func memcachedLeaderTest(t *testing.T, f *framework.Framework, ctx *framework.TestCtx) error {
+func memcachedLeaderTest(t *testing.T, f *framework.Framework, ctx *framework.Context) error {
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
 		return err
@@ -160,7 +160,7 @@ func verifyLeader(t *testing.T, namespace string, f *framework.Framework, labels
 	return nil, fmt.Errorf("did not find operator pod that was referenced by configmap")
 }
 
-func memcachedScaleTest(t *testing.T, f *framework.Framework, ctx *framework.TestCtx, fromReplicas, toReplicas int) error {
+func memcachedScaleTest(t *testing.T, f *framework.Framework, ctx *framework.Context, fromReplicas, toReplicas int) error {
 	name := "example-memcached"
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -177,7 +177,7 @@ func memcachedScaleTest(t *testing.T, f *framework.Framework, ctx *framework.Tes
 			Size: int32(fromReplicas),
 		},
 	}
-	// use TestCtx's create helper to create the object and add a cleanup function for the new object
+	// use Context's create helper to create the object and add a cleanup function for the new object
 	err = f.Client.Create(goctx.TODO(), exampleMemcached, &framework.CleanupOptions{TestContext: ctx, Timeout: cleanupTimeout, RetryInterval: cleanupRetryInterval})
 	if err != nil {
 		return fmt.Errorf("could no create CR: %v", err)
@@ -211,7 +211,7 @@ func memcachedScaleTest(t *testing.T, f *framework.Framework, ctx *framework.Tes
 
 func MemcachedLocal(t *testing.T) {
 	// get global framework variables
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
@@ -265,7 +265,7 @@ func MemcachedLocal(t *testing.T) {
 
 func MemcachedCluster(t *testing.T) {
 	// get global framework variables
-	ctx := framework.NewTestCtx(t)
+	ctx := framework.NewContext(t)
 	defer ctx.Cleanup()
 
 	err := ctx.InitializeClusterResources(&framework.CleanupOptions{TestContext: ctx, Timeout: cleanupTimeout, RetryInterval: cleanupRetryInterval})
@@ -305,7 +305,7 @@ func MemcachedCluster(t *testing.T) {
 	t.Log("Completed memcached custom resource metrics test")
 }
 
-func memcachedMetricsTest(t *testing.T, f *framework.Framework, ctx *framework.TestCtx) error {
+func memcachedMetricsTest(t *testing.T, f *framework.Framework, ctx *framework.Context) error {
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
 		return err
@@ -347,7 +347,7 @@ func memcachedMetricsTest(t *testing.T, f *framework.Framework, ctx *framework.T
 	return nil
 }
 
-func memcachedOperatorMetricsTest(t *testing.T, f *framework.Framework, ctx *framework.TestCtx) error {
+func memcachedOperatorMetricsTest(t *testing.T, f *framework.Framework, ctx *framework.Context) error {
 	namespace, err := ctx.GetNamespace()
 	if err != nil {
 		return err